Organizacion y cambios de Helpers
